+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Adding `validateSource` to `HtmlValidate` allowing to manually passing a
   source.
+- `HtmlValidate.getConfigFor` is now part of public API.
 
 ## 0.16.1 (2018-12-16)